@@ -1,16 +1,13 @@
 package ui.activity;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.app.AlertDialog;
 import android.content.DialogInterface;
 import android.os.Bundle;
-import android.support.v4.widget.SwipeRefreshLayout;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 import android.text.InputType;
 import android.view.LayoutInflater;
 import android.view.View;
-import android.support.v4.app.Fragment;
-
-import android.support.v7.widget.Toolbar;
 
 import android.view.ViewGroup;
 import android.widget.AdapterView;
